Refactor some of the app initialization code
- Move SessionUtils into jsapi-utils package
- Move PluginUtils and some of the app loading stuff into app-utils package
