+++ packages/app-utils/package.json
@@ -1,14 +1,14 @@
 {
-  "name": "@deephaven/plugin-utils",
+  "name": "@deephaven/app-utils",
   "version": "0.35.0",
-  "description": "Deephaven Plugin Utils",
+  "description": "Deephaven App Utils",
   "author": "Deephaven Data Labs LLC",
   "license": "Apache-2.0",
   "type": "module",
   "repository": {
     "type": "git",
     "url": "https://github.com/deephaven/web-client-ui.git",
-    "directory": "packages/plugin-utils"
+    "directory": "packages/app-utils"
   },
   "source": "src/index.js",
   "main": "dist/index.js",
@@ -28,6 +28,9 @@
     "redux": "^4.x"
   },
   "dependencies": {
+    "@deephaven/auth-plugin": "file:../auth-plugin",
+    "@deephaven/auth-core-plugins": "file:../auth-core-plugins",
+    "@deephaven/jsapi-types": "file:../jsapi-types",
     "@deephaven/log": "file:../log",
     "@paciolan/remote-component": "2.13.0",
     "@paciolan/remote-module-loader": "^3.0.2"
