+++ packages/app-utils/src/remote-component.config.js
