+++ packages/code-studio/package.json
@@ -29,7 +29,7 @@
     "@deephaven/jsapi-types": "file:../jsapi-types",
     "@deephaven/jsapi-utils": "file:../jsapi-utils",
     "@deephaven/log": "file:../log",
-    "@deephaven/plugin-utils": "file:../plugin-utils",
+    "@deephaven/app-utils": "file:../app-utils",
     "@deephaven/pouch-storage": "file:../pouch-storage",
     "@deephaven/react-hooks": "file:../react-hooks",
     "@deephaven/redux": "file:../redux",
