+++ packages/code-studio/tsconfig.json
@@ -25,7 +25,7 @@
     { "path": "../jsapi-types" },
     { "path": "../jsapi-utils" },
     { "path": "../log" },
-    { "path": "../plugin-utils" },
+    { "path": "../app-utils" },
     { "path": "../pouch-storage" },
     { "path": "../react-hooks" },
     { "path": "../redux" },
