+++ packages/dashboard-core-plugins/src/panels/ConsolePanel.test.tsx
@@ -3,8 +3,8 @@ import { render } from '@testing-library/react';
 import { CommandHistoryStorage } from '@deephaven/console';
 import type { Container } from '@deephaven/golden-layout';
 import { IdeConnection, IdeSession } from '@deephaven/jsapi-types';
+import { SessionConfig, SessionWrapper } from '@deephaven/jsapi-utils';
 import { ConsolePanel } from './ConsolePanel';
-import { SessionConfig, SessionWrapper } from '../redux';
 
 const mockConsole = jest.fn(() => null);
 jest.mock('@deephaven/console', () => ({
