+++ packages/jsapi-utils/src/NoConsolesError.ts
