+++ packages/jsapi-utils/src/index.ts
@@ -5,5 +5,7 @@ export * from './Formatter';
 export { default as FormatterUtils } from './FormatterUtils';
 export * from './FormatterUtils';
 export * from './MessageUtils';
+export * from './NoConsolesError';
+export * from './SessionUtils';
 export * from './Settings';
 export * from './TableUtils';
