+++ packages/redux/package.json
@@ -24,7 +24,7 @@
   "dependencies": {
     "@deephaven/jsapi-utils": "file:../jsapi-utils",
     "@deephaven/log": "file:../log",
-    "@deephaven/plugin-utils": "file:../plugin-utils",
+    "@deephaven/app-utils": "file:../app-utils",
     "deep-equal": "^2.0.5",
     "redux-thunk": "2.4.1"
   },
